authservice/requirements.txt
Django==3.2.9
djangorestframework_simplejwt
requests
django-cors-headers